Bindings improvements for the Februrary 2021 update (#411)
* Native handles, Veldrid support, and other useful enhancements

* Regenerate bindings, support for constants and macro constants

* Fix some really embarassing typing bugs, start D3DTriangle

* More typing bug fixes, update to latest Vulkan

* wip

* Add $CPPRETFIXUP intrinsic

* Move the now working D3D12Triangle to the lab

* Ref-returning fused properties on .NET Standard 2.1 platforms

* Delete gl.json

* Delete glcore.json

* Switch back to 32-bit mode

* Fix calling convention detection

* Respect native api calling convention

* Alter NativeApiAttribute slightly

* OpenGL token revival (fixes #400)

* Laziness improvements, hopefully fixing #407

* Update src/Core/Silk.NET.Core/Native/ComPtr.cs

+++ src/Assimp/Silk.NET.Assimp/Structs/PfnFileReadProc.gen.cs
@@ -21,10 +21,10 @@ namespace Silk.NET.Assimp
     public unsafe readonly struct PfnFileReadProc : IDisposable
     {
         private readonly void* _handle;
-        public delegate* unmanaged[Cdecl]<File*, byte*, uint, uint, uint> Handle => (delegate* unmanaged[Cdecl]<File*, byte*, uint, uint, uint>) _handle;
+        public delegate* unmanaged[Cdecl]<File*, byte*, nuint, nuint, nuint> Handle => (delegate* unmanaged[Cdecl]<File*, byte*, nuint, nuint, nuint>) _handle;
         public PfnFileReadProc
         (
-            delegate* unmanaged[Cdecl]<File*, byte*, uint, uint, uint> ptr
+            delegate* unmanaged[Cdecl]<File*, byte*, nuint, nuint, nuint> ptr
         ) => _handle = ptr;
 
         public PfnFileReadProc
@@ -37,7 +37,7 @@ namespace Silk.NET.Assimp
 
         public static implicit operator nint(PfnFileReadProc pfn) => (nint) pfn.Handle;
         public static explicit operator PfnFileReadProc(nint pfn)
-            => new PfnFileReadProc((delegate* unmanaged[Cdecl]<File*, byte*, uint, uint, uint>) pfn);
+            => new PfnFileReadProc((delegate* unmanaged[Cdecl]<File*, byte*, nuint, nuint, nuint>) pfn);
 
         public static implicit operator PfnFileReadProc(FileReadProc proc)
             => new PfnFileReadProc(proc);
@@ -45,10 +45,10 @@ namespace Silk.NET.Assimp
         public static explicit operator FileReadProc(PfnFileReadProc pfn)
             => SilkMarshal.PtrToDelegate<FileReadProc>(pfn);
 
-        public static implicit operator delegate* unmanaged[Cdecl]<File*, byte*, uint, uint, uint>(PfnFileReadProc pfn) => pfn.Handle;
-        public static implicit operator PfnFileReadProc(delegate* unmanaged[Cdecl]<File*, byte*, uint, uint, uint> ptr) => new PfnFileReadProc(ptr);
+        public static implicit operator delegate* unmanaged[Cdecl]<File*, byte*, nuint, nuint, nuint>(PfnFileReadProc pfn) => pfn.Handle;
+        public static implicit operator PfnFileReadProc(delegate* unmanaged[Cdecl]<File*, byte*, nuint, nuint, nuint> ptr) => new PfnFileReadProc(ptr);
     }
 
-    public unsafe delegate uint FileReadProc(File* arg0, byte* arg1, uint arg2, uint arg3);
+    public unsafe delegate nuint FileReadProc(File* arg0, byte* arg1, nuint arg2, nuint arg3);
 }
