+++ src/Microsoft/Silk.NET.Direct3D12/Structs/VideoProcessTransform.gen.cs
@@ -23,8 +23,8 @@ namespace Silk.NET.Direct3D12
     {
         public VideoProcessTransform
         (
-            Silk.NET.Maths.Rectangle<long>? sourceRectangle = null,
-            Silk.NET.Maths.Rectangle<long>? destinationRectangle = null,
+            Silk.NET.Maths.Rectangle<int>? sourceRectangle = null,
+            Silk.NET.Maths.Rectangle<int>? destinationRectangle = null,
             VideoProcessOrientation? orientation = null
         ) : this()
         {
@@ -48,12 +48,12 @@ namespace Silk.NET.Direct3D12
         [NativeName("Type", "D3D12_RECT")]
         [NativeName("Type.Name", "D3D12_RECT")]
         [NativeName("Name", "SourceRectangle")]
-        public Silk.NET.Maths.Rectangle<long> SourceRectangle;
+        public Silk.NET.Maths.Rectangle<int> SourceRectangle;
 
         [NativeName("Type", "D3D12_RECT")]
         [NativeName("Type.Name", "D3D12_RECT")]
         [NativeName("Name", "DestinationRectangle")]
-        public Silk.NET.Maths.Rectangle<long> DestinationRectangle;
+        public Silk.NET.Maths.Rectangle<int> DestinationRectangle;
 
         [NativeName("Type", "D3D12_VIDEO_PROCESS_ORIENTATION")]
         [NativeName("Type.Name", "D3D12_VIDEO_PROCESS_ORIENTATION")]
